the actual files and stuff
the designer is used to design the ui duh
to use the ui you designed, add the uiclass and helper funcs along with the text file that stores your ui and the font you chose into your target projects folder (where the actual .pde file is located), then run the load function in setup with the needed addresses and run updateAll() in draw. to access the values of the ui elements, just access them like any object in an arraylist and either look at the relevent values or run the correct functions. I have included an example ui design and some fonts